Ensures that the Dqlite node ID (uint64) is stored as text in Dqlite, to avoid driver issues converting to int.
We can still scan it straight into uint64, so the change is minimal.

+++ domain/controllernode/state/state_test.go
@@ -59,13 +59,13 @@ func (s *stateSuite) TestUpdateUpdateDqliteNode(c *gc.C) {
 	c.Assert(row.Err(), jc.ErrorIsNil)
 
 	var (
-		id   int
+		id   uint64
 		addr string
 	)
 	err = row.Scan(&id, &addr)
 	c.Assert(err, jc.ErrorIsNil)
 
-	c.Check(id, gc.Equals, 12345)
+	c.Check(id, gc.Equals, uint64(12345))
 	c.Check(addr, gc.Equals, "192.168.5.60")
 }
